Extract `resolveCountryCodes` from `useBudgetRecommendation` hook as `useCountryCodesForBudgetQuery` hook.
